feat(gltf_auto_export):  * found source of export issues ! mismatch between context.scene & context.window.scene (the last we can set, the other not)  * overhauled & cleaned up materials handling to avoid context issues & reliance on bpy.ops  * cleanups  * added bug detector (wip) & started experiment with temp_override
