--- tools/gltf_auto_export/__init__alt.py
@@ -1,166 +1,0 @@
-bl_info = {
-    "name": "gltf_auto_export",
-    "author": "kaosigh",
-    "version": (0, 10, 0),
-    "blender": (3, 4, 0),
-    "location": "File > Import-Export",
-    "description": "glTF/glb auto-export",
-    "warning": "",
-    "wiki_url": "https://github.com/kaosat-dev/Blender_bevy_components_workflow",
-    "tracker_url": "https://github.com/kaosat-dev/Blender_bevy_components_workflow/issues/new",
-    "category": "Import-Export"
-}
-
-import bpy
-from bpy.props import (BoolProperty,
-                       IntProperty,
-                       StringProperty,
-                       EnumProperty,
-                       CollectionProperty
-                       )
-
-
-# glTF extensions are named following a convention with known prefixes.
-# See: https://github.com/KhronosGroup/glTF/tree/main/extensions#about-gltf-extensions
-# also: https://github.com/KhronosGroup/glTF/blob/main/extensions/Prefixes.md
-glTF_extension_name = "EXT_auto_export"
-
-# Support for an extension is "required" if a typical glTF viewer cannot be expected
-# to load a given model without understanding the contents of the extension.
-# For example, a compression scheme or new image format (with no fallback included)
-# would be "required", but physics metadata or app-specific settings could be optional.
-extension_is_required = False
-
-class ExampleExtensionProperties(bpy.types.PropertyGroup):
-    enabled: bpy.props.BoolProperty(
-        name=bl_info["name"],
-        description='Include this extension in the exported glTF file.',
-        default=True
-        )
-    
-    auto_export_main_scene_name: StringProperty(
-        name='Main scene',
-        description='The name of the main scene/level/world to auto export',
-        default='Scene'
-    )
-    auto_export_output_folder: StringProperty(
-        name='Export folder (relative)',
-        description='The root folder for all exports(relative to current file) Defaults to current folder',
-        default=''
-    )
-    auto_export_library_scene_name: StringProperty(
-        name='Library scene',
-        description='The name of the library scene to auto export',
-        default='Library'
-    )
-    # scene components
-    auto_export_scene_settings: BoolProperty(
-        name='Export scene settings',
-        description='Export scene settings ie AmbientLighting, Bloom, AO etc',
-        default=False
-    )
-
-    # blueprint settings
-    auto_export_blueprints: BoolProperty(
-        name='Export Blueprints',
-        description='Replaces collection instances with an Empty with a BlueprintName custom property',
-        default=True
-    )
-    auto_export_blueprints_path: StringProperty(
-        name='Blueprints path',
-        description='path to export the blueprints to (relative to the Export folder)',
-        default='library'
-    )
-
-    auto_export_materials_library: BoolProperty(
-        name='Export materials library',
-        description='remove materials from blueprints and use the material library instead',
-        default=False
-    )
-    auto_export_materials_path: StringProperty(
-        name='Materials path',
-        description='path to export the materials libraries to (relative to the root folder)',
-        default='materials'
-    )
-
-def register():
-    bpy.utils.register_class(ExampleExtensionProperties)
-    bpy.types.Scene.ExampleExtensionProperties = bpy.props.PointerProperty(type=ExampleExtensionProperties)
-
-def register_panel():
-    # Register the panel on demand, we need to be sure to only register it once
-    # This is necessary because the panel is a child of the extensions panel,
-    # which may not be registered when we try to register this extension
-    try:
-        bpy.utils.register_class(GLTF_PT_UserExtensionPanel)
-    except Exception:
-        pass
-
-    # If the glTF exporter is disabled, we need to unregister the extension panel
-    # Just return a function to the exporter so it can unregister the panel
-    return unregister_panel
-
-
-def unregister_panel():
-    # Since panel is registered on demand, it is possible it is not registered
-    try:
-        bpy.utils.unregister_class(GLTF_PT_UserExtensionPanel)
-    except Exception:
-        pass
-
-
-def unregister():
-    unregister_panel()
-    bpy.utils.unregister_class(ExampleExtensionProperties)
-    del bpy.types.Scene.ExampleExtensionProperties
-
-class GLTF_PT_UserExtensionPanel(bpy.types.Panel):
-
-    bl_space_type = 'FILE_BROWSER'
-    bl_region_type = 'TOOL_PROPS'
-    bl_label = "Enabled"
-    bl_parent_id = "GLTF_PT_export_user_extensions"
-    bl_options = {'DEFAULT_CLOSED'}
-
-    @classmethod
-    def poll(cls, context):
-        sfile = context.space_data
-        operator = sfile.active_operator
-        return operator.bl_idname == "EXPORT_SCENE_OT_gltf"
-
-    def draw_header(self, context):
-        props = bpy.context.scene.ExampleExtensionProperties
-        self.layout.prop(props, 'enabled')
-
-    def draw(self, context):
-        layout = self.layout
-        layout.use_property_split = True
-        layout.use_property_decorate = False  # No animation.
-
-        props = bpy.context.scene.ExampleExtensionProperties
-        layout.active = props.enabled
-
-        props = bpy.context.scene.ExampleExtensionProperties
-        for bla in props.__annotations__:
-            layout.prop(props, bla)
-
-
-class glTF2ExportUserExtension:
-
-    def __init__(self):
-        # We need to wait until we create the gltf2UserExtension to import the gltf2 modules
-        # Otherwise, it may fail because the gltf2 may not be loaded yet
-        from io_scene_gltf2.io.com.gltf2_io_extensions import Extension
-        self.Extension = Extension
-        self.properties = bpy.context.scene.ExampleExtensionProperties
-
-    def gather_node_hook(self, gltf2_object, blender_object, export_settings):
-        if self.properties.enabled:
-            if gltf2_object.extensions is None:
-                gltf2_object.extensions = {}
-            print("bla bla")
-            gltf2_object.extensions[glTF_extension_name] = self.Extension(
-                name=glTF_extension_name,
-                extension={"auto_export_blueprints": self.properties.auto_export_blueprints},
-                required=extension_is_required
-            )
