+++ tools/gltf_auto_export/bug_detector.py
@@ -1,0 +1,4 @@
+import bpy
+
+def detect_context_mismatch():
+    return bpy.context.scene == bpy.context.window.scene
