+++ tools/gltf_auto_export/materials.py
@@ -1,6 +1,7 @@
 
+from .helpers_collections import set_active_collection
 from .helpers_export import export_gltf, generate_gltf_export_preferences
-from .helpers import traverse_tree
+from .helpers import make_cube, traverse_tree
 import bpy
 import os
 from pathlib import Path
@@ -46,32 +47,33 @@ def get_all_materials(collection_names, library_scenes):
 
 
 # creates a new object with the applied material, for the material library
-def make_material_object(name, location, rotation, scale, material): 
-    original_active_object = bpy.context.active_object
-    # bpy.ops.object.empty_add(type='PLAIN_AXES', location=location, rotation=rotation, scale=scale)
-    bpy.ops.mesh.primitive_cube_add(size=0.1, location=location)  
-    object = bpy.context.active_object
-    object.name = name
-    #obj.scale = scale # scale is not set correctly ?????
-    if object.data.materials:
-        # assign to 1st material slot
-        object.data.materials[0] = material
-    else:
-        # no slots
-        object.data.materials.append(material)
+def make_material_object(name, location=[0,0,0], rotation=[0,0,0], scale=[1,1,1], material=None, collection=None): 
+    #original_active_object = bpy.context.active_object
+    #bpy.ops.mesh.primitive_cube_add(size=0.1, location=location)  
+    object = make_cube(name, location=location, rotation=rotation, scale=scale, collection=collection)
+    if material:
+        if object.data.materials:
+            # assign to 1st material slot
+            object.data.materials[0] = material
+        else:
+            # no slots
+            object.data.materials.append(material)
 
-    bpy.context.view_layer.objects.active = original_active_object
+    #bpy.context.view_layer.objects.active = original_active_object
     return object
 
 
 # generates a materials scene: 
 def generate_materials_scenes(used_material_names):
+    #bpy.ops.scene.new(type='NEW')#name="__materials_scene")
+    #bpy.context.scene.name = "__materials_scene"
+    #temp_scene = bpy.data.scenes["__materials_scene"]
     temp_scene = bpy.data.scenes.new(name="__materials_scene")
-    bpy.context.window.scene = temp_scene
+    root_collection = temp_scene.collection
 
     for index, material_name in enumerate(used_material_names):
         material = bpy.data.materials[material_name]
-        make_material_object("Material_"+material_name, [index * 0.2,0,0], [], [], material)
+        make_material_object("Material_"+material_name, [index * 0.2,0,0], material=material, collection=root_collection)
 
     # we set our active scene to be this one : this is needed otherwise the stand-in objects get generated in the wrong scene
     return temp_scene
@@ -80,7 +82,18 @@ def clear_materials_scene(temp_scene):
     root_collection = temp_scene.collection 
     scene_objects = [o for o in root_collection.objects]
     for object in scene_objects:
-        bpy.data.objects.remove(object, do_unlink=True)
+        #print("removing ", object)
+        try:
+            mesh = bpy.data.meshes[object.name+"_Mesh"]
+            bpy.data.meshes.remove(mesh, do_unlink=True)
+        except Exception as error:
+            pass
+            #print("could not remove mesh", error)
+            
+        try:
+            bpy.data.objects.remove(object, do_unlink=True)
+        except:pass
+
     bpy.data.scenes.remove(temp_scene)
 
 # exports the materials used inside the current project:
@@ -92,12 +105,18 @@ def export_materials(collections, library_scenes, folder_path, addon_prefs):
     used_material_names = get_all_materials(collections, library_scenes)
     current_project_name = Path(bpy.context.blend_data.filepath).stem
 
-    print("materials", used_material_names)
-
     # save the current active scene
     current_scene = bpy.context.window.scene
-    mat_scene = generate_materials_scenes(used_material_names)
+    # and the active collection
+    active_collection =  bpy.context.view_layer.active_layer_collection
 
+    # now do the real work
+    # ----------------------------------------------
+    mat_scene = generate_materials_scenes(used_material_names)
+    #set_active_collection(mat_scene, "Scene Collection")
+    bpy.context.window.scene = mat_scene
+    #print("active scene", bpy.context.window.scene.name, bpy.context.scene.name)
+    #print("scene infos", mat_scene.objects)
 
     gltf_output_path = os.path.join(folder_path, export_materials_path, current_project_name + "_materials_library")
     print("       exporting Materials to", gltf_output_path, ".gltf/glb")
@@ -110,12 +129,14 @@ def export_materials(collections, library_scenes, folder_path, addon_prefs):
                     'export_apply':True
                     }
     export_gltf(gltf_output_path, export_settings)
-
+    # -----------------------------------------------
     # remove materials scenes
     clear_materials_scene(mat_scene)
-
     # reset scene to previously selected scene
     bpy.context.window.scene = current_scene
+    # reset active collection
+    bpy.context.view_layer.active_layer_collection = active_collection
+
 
 def cleanup_materials(collections, library_scenes):
     # remove temporary components
